fix(chat): complete delimiter parsing
